StudiFlowApi/requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.31.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
effdet==0.4.1
emoji==2.12.1
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
iopath==0.1.10
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.2.3
langchain-core==0.2.33
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.99
layoutparser==0.3.4
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
openai==1.41.0
opencv-python==4.10.0.84
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.3
pikepdf==9.1.1
pillow==10.4.0
pillow_heif==0.18.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.4
scipy==1.14.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.0
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.14.4
unstructured-client==0.25.5
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.13
urllib3==2.2.2
whitenoise==6.7.0
wrapt==1.16.0
yarl==1.9.4
